requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.28.0

# Healthcare data processing
pydicom>=2.3.0
hl7apy>=1.3.0
python-dateutil>=2.8.0

# Data validation and quality
pandera>=0.15.0
great-expectations>=0.17.0
scipy>=1.10.0

# Statistical analysis
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Database connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and logging
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# API and web services
fastapi>=0.100.0
uvicorn>=0.23.0